requirements.txt
sayn
numpy
pandas
matplotlib
Pillow
wordcloud
nltk
